optimized package structure
