+++ src/main/java/sql/dao/IManufacturerDAO.java
@@ -1,4 +1,4 @@
-package sql.dao.entity.manufacturer;
+package sql.dao;
 
 import sql.dao.IBaseDAO;
 import sql.datamodels.entity.Manufacturer;
