+++ src/main/java/sql/dao/mysql/CountryDaoMySQL.java
@@ -1,37 +1,36 @@
-package sql.dao.entity.city;
+package sql.dao.mysql;
 
-import sql.dao.MySQLDAO;
-import sql.dao.MySQLType;
-import sql.datamodels.entity.City;
+import sql.dao.ICountryDAO;
+import sql.datamodels.entity.Country;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class CityDaoMySQL extends MySQLDAO implements ICityDAO {
+public class CountryDaoMySQL extends MySQLDAO implements ICountryDAO {
 
     @Override
     public Optional get(int id) {
-        //public City(String name)
-        String sql = "SELECT name FROM cities WHERE id_city = ?";
-        return Optional.ofNullable(new City(getWithTryCatch(sql, id)));
+        //public Country(String name)
+        String sql = "SELECT name FROM countries WHERE id_country = ?";
+        return Optional.ofNullable(new Country(getWithTryCatch(sql, id)));
     }
 
     @Override
     public List getAll() {
-        //public City(String name)
-        String sql = "SELECT name FROM cities";
+        //public Country(String name)
+        String sql = "SELECT name FROM countries";
         List<Object[]> list = getWithTryCatch(sql);
-        List<City> output = new ArrayList<>();
+        List<Country> output = new ArrayList<>();
         for (Object[] o : list) {
-            output.add(new City(o));
+            output.add(new Country(o));
         }
         return output;
     }
 
     @Override
-    public void save(City o) {
-        String sql = "INSERT INTO cities(name) VALUES (?)";
+    public void save(Country o) {
+        String sql = "INSERT INTO countries(name) VALUES (?)";
         List<Object> valueList = new ArrayList<>();
         valueList.add(o.getName());
 
@@ -42,8 +41,8 @@ public class CityDaoMySQL extends MySQLDAO implements ICityDAO {
     }
 
     @Override
-    public void update(City o, int id) {
-        String sql = "UPDATE cities SET name = ? WHERE id_city = ?";
+    public void update(Country o, int id) {
+        String sql = "UPDATE countries SET name = ? WHERE id_country = ?";
         List<Object> valueList = new ArrayList<>();
         valueList.add(o.getName());
         valueList.add(id);
@@ -57,7 +56,7 @@ public class CityDaoMySQL extends MySQLDAO implements ICityDAO {
 
     @Override
     public void delete(int id) {
-        String sql = "DELETE FROM cities WHERE id_city = ?";
+        String sql = "DELETE FROM countries WHERE id_country = ?";
         deleteWithTryCatch(sql, id);
     }
 }
